base,misc: Collapse and eliminate the ULL and LL macros.
These just move the ULL or LL suffix to the value in question, and cast
to a uint64_t or an int64_t. We should be able to drop the cast
entirely, and turn the macro into a suffix for the literals in question.

Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/42505
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: kokoro <noreply+kokoro@google.com>

+++ src/arch/arm/interrupts.hh
@@ -93,7 +93,7 @@ class Interrupts : public BaseInterrupts
             panic("No support for other interrupt indexes\n");
 
         interrupts[int_num] = true;
-        intStatus |= ULL(1) << int_num;
+        intStatus |= 1ULL << int_num;
     }
 
     void
@@ -108,7 +108,7 @@ class Interrupts : public BaseInterrupts
             panic("No support for other interrupt indexes\n");
 
         interrupts[int_num] = false;
-        intStatus &= ~(ULL(1) << int_num);
+        intStatus &= ~(1ULL << int_num);
     }
 
     void
