+++ src/arch/arm/isa_traits.hh
@@ -50,7 +50,7 @@ namespace ArmISA
     const ByteOrder GuestByteOrder = ByteOrder::little;
 
     const Addr PageShift = 12;
-    const Addr PageBytes = ULL(1) << PageShift;
+    const Addr PageBytes = 1ULL << PageShift;
 } // namespace ArmISA
 
 #endif // __ARCH_ARM_ISA_TRAITS_HH__
