+++ src/arch/x86/isa_traits.hh
@@ -46,7 +46,7 @@ namespace X86ISA
     const ByteOrder GuestByteOrder = ByteOrder::little;
 
     const Addr PageShift = 12;
-    const Addr PageBytes = ULL(1) << PageShift;
+    const Addr PageBytes = 1ULL << PageShift;
 }
 
 #endif // __ARCH_X86_ISATRAITS_HH__
