+++ src/cpu/pred/tage_base.cc
@@ -128,7 +128,7 @@ TAGEBase::init()
         initFoldedHistories(history);
     }
 
-    const uint64_t bimodalTableSize = ULL(1) << logTagTableSizes[0];
+    const uint64_t bimodalTableSize = 1ULL << logTagTableSizes[0];
     btablePrediction.resize(bimodalTableSize, false);
     btableHysteresis.resize(bimodalTableSize >> logRatioBiModalHystEntries,
                             true);
@@ -200,7 +200,7 @@ TAGEBase::btbUpdate(ThreadID tid, Addr branch_pc, BranchInfo* &bi)
 int
 TAGEBase::bindex(Addr pc_in) const
 {
-    return ((pc_in >> instShiftAmt) & ((ULL(1) << (logTagTableSizes[0])) - 1));
+    return ((pc_in >> instShiftAmt) & ((1ULL << (logTagTableSizes[0])) - 1));
 }
 
 int
@@ -208,13 +208,13 @@ TAGEBase::F(int A, int size, int bank) const
 {
     int A1, A2;
 
-    A = A & ((ULL(1) << size) - 1);
-    A1 = (A & ((ULL(1) << logTagTableSizes[bank]) - 1));
+    A = A & ((1ULL << size) - 1);
+    A1 = (A & ((1ULL << logTagTableSizes[bank]) - 1));
     A2 = (A >> logTagTableSizes[bank]);
-    A2 = ((A2 << bank) & ((ULL(1) << logTagTableSizes[bank]) - 1))
+    A2 = ((A2 << bank) & ((1ULL << logTagTableSizes[bank]) - 1))
        + (A2 >> (logTagTableSizes[bank] - bank));
     A = A1 ^ A2;
-    A = ((A << bank) & ((ULL(1) << logTagTableSizes[bank]) - 1))
+    A = ((A << bank) & ((1ULL << logTagTableSizes[bank]) - 1))
       + (A >> (logTagTableSizes[bank] - bank));
     return (A);
 }
@@ -233,7 +233,7 @@ TAGEBase::gindex(ThreadID tid, Addr pc, int bank) const
         threadHistory[tid].computeIndices[bank].comp ^
         F(threadHistory[tid].pathHist, hlen, bank);
 
-    return (index & ((ULL(1) << (logTagTableSizes[bank])) - 1));
+    return (index & ((1ULL << (logTagTableSizes[bank])) - 1));
 }
 
 
@@ -245,7 +245,7 @@ TAGEBase::gtag(ThreadID tid, Addr pc, int bank) const
               threadHistory[tid].computeTags[0][bank].comp ^
               (threadHistory[tid].computeTags[1][bank].comp << 1);
 
-    return (tag & ((ULL(1) << tagTableTagWidths[bank]) - 1));
+    return (tag & ((1ULL << tagTableTagWidths[bank]) - 1));
 }
 
 
@@ -451,7 +451,7 @@ TAGEBase::handleAllocAndUReset(bool alloc, bool taken, BranchInfo* bi,
         // to  avoid ping-pong, we do not choose systematically the next
         // entry, but among the 3 next entries
         int Y = nrand &
-            ((ULL(1) << (nHistoryTables - bi->hitBank - 1)) - 1);
+            ((1ULL << (nHistoryTables - bi->hitBank - 1)) - 1);
         int X = bi->hitBank + 1;
         if (Y & 1) {
             X++;
@@ -487,11 +487,11 @@ void
 TAGEBase::handleUReset()
 {
     //periodic reset of u: reset is not complete but bit by bit
-    if ((tCounter & ((ULL(1) << logUResetPeriod) - 1)) == 0) {
+    if ((tCounter & ((1ULL << logUResetPeriod) - 1)) == 0) {
         // reset least significant bit
         // most significant bit becomes least significant bit
         for (int i = 1; i <= nHistoryTables; i++) {
-            for (int j = 0; j < (ULL(1) << logTagTableSizes[i]); j++) {
+            for (int j = 0; j < (1ULL << logTagTableSizes[i]); j++) {
                 resetUctr(gtable[i][j].u);
             }
         }
@@ -593,7 +593,7 @@ TAGEBase::updateHistories(ThreadID tid, Addr branch_pc, bool taken,
     //update user history
     updateGHist(tHist.gHist, taken, tHist.globalHistory, tHist.ptGhist);
     tHist.pathHist = (tHist.pathHist << 1) + pathbit;
-    tHist.pathHist = (tHist.pathHist & ((ULL(1) << pathHistBits) - 1));
+    tHist.pathHist = (tHist.pathHist & ((1ULL << pathHistBits) - 1));
 
     if (speculative) {
         bi->ptGhist = tHist.ptGhist;
@@ -788,7 +788,7 @@ TAGEBase::getSizeInBits() const {
         bits += (1 << logTagTableSizes[i]) *
             (tagTableCounterBits + tagTableUBits + tagTableTagWidths[i]);
     }
-    uint64_t bimodalTableSize = ULL(1) << logTagTableSizes[0];
+    uint64_t bimodalTableSize = 1ULL << logTagTableSizes[0];
     bits += numUseAltOnNa * useAltOnNaBits;
     bits += bimodalTableSize;
     bits += (bimodalTableSize >> logRatioBiModalHystEntries);
