+++ src/mem/mem_interface.cc
@@ -775,7 +775,7 @@ DRAMInterface::DRAMInterface(const DRAMInterfaceParams &_p)
     uint64_t deviceCapacity = deviceSize / (1024 * 1024) * devicesPerRank *
                               ranksPerChannel;
 
-    uint64_t capacity = ULL(1) << ceilLog2(AbstractMemory::size());
+    uint64_t capacity = 1ULL << ceilLog2(AbstractMemory::size());
 
     DPRINTF(DRAM, "Memory capacity %lld (%lld) bytes\n", capacity,
             AbstractMemory::size());
@@ -2038,7 +2038,7 @@ NVMInterface::NVMInterface(const NVMInterfaceParams &_p)
         ranks.push_back(rank);
     }
 
-    uint64_t capacity = ULL(1) << ceilLog2(AbstractMemory::size());
+    uint64_t capacity = 1ULL << ceilLog2(AbstractMemory::size());
 
     DPRINTF(NVM, "NVM capacity %lld (%lld) bytes\n", capacity,
             AbstractMemory::size());
